# Optimize `get_unfilled_outstanding_batches()` query